# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Management API",
        "description": "\n *     ## Инструкция по использованию API:\n *     \n *     1. **Регистрация/Вход**: Сначала зарегистрируйтесь через `/register` или войдите через `/login`\n *     2. **Получите токен**: После успешной аутентификации вы получите токен в ответе\n *     3. **Авторизация**: Нажмите кнопку 'Authorize' и введите: `Bearer ваш_токен`\n *     4. **Тестирование**: Теперь можете тестировать защищенные endpoints\n *     \n *     ## Публичные endpoints (не требуют авторизации):\n *     - GET /tasks-all - Все задачи\n *     - GET /tasks/{id} - Конкретная задача\n *     - GET /tasks-user/{id} - Задачи пользователя\n *     \n *     ## Защищенные endpoints (требуют авторизации):\n *     - POST /tasks - Создать задачу\n *     - PATCH /tasks/{id} - Обновить задачу\n *     - DELETE /tasks/{id} - Удалить задачу\n *     - POST /logout - Выход\n *     - GET /user - Информация о пользователе",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация нового пользователя",
                "description": "Создает нового пользователя и возвращает токен для аутентификации. Скопируйте полученный токен и используйте его для авторизации в защищенных endpoints.",
                "operationId": "21cce9fb60fd5fd9d235c8a9b2f477cc",
                "requestBody": {
                    "description": "Данные для регистрации",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Имя пользователя",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Email пользователя",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "description": "Пароль (мин. 8 символов)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "description": "Подтверждение пароля",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешная регистрация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "description": "Токен для аутентификации. Скопируйте этот токен и используйте для авторизации.",
                                            "type": "string",
                                            "example": "1|abcdef1234567890"
                                        },
                                        "token_type": {
                                            "description": "Тип токена",
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The email has already been taken."
                                                ],
                                                "password": [
                                                    "The password confirmation does not match."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Вход в систему",
                "description": "Аутентификация пользователя и получение токена. Используйте этот токен для доступа к защищенным endpoints.",
                "operationId": "614c182a075050a98da662867ad0229b",
                "requestBody": {
                    "description": "Учетные данные для входа",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email пользователя",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "description": "Пароль пользователя",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный вход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "description": "Токен аутентификации. Скопируйте этот токен для авторизации.",
                                            "type": "string",
                                            "example": "1|abcdef1234567890"
                                        },
                                        "token_type": {
                                            "description": "Тип токена",
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The provided credentials are incorrect."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Выход из системы",
                "description": "Удаляет текущий токен аутентификации пользователя.",
                "operationId": "666f1aa6a0e7631cc0017c6bc97cf9f3",
                "responses": {
                    "200": {
                        "description": "Успешный выход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Информация о текущем пользователе",
                "description": "Возвращает информацию об аутентифицированном пользователе.",
                "operationId": "654bd2e8f33d21e8dcadf0464a6b5e5b",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Аутентификация через Sanctum Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}